+++ R/autoplot-multi-dist.R
@@ -227,7 +227,7 @@ tidy_multi_dist_autoplot <- function(.data, .plot_type = "density", .line_size =
       ggplot2::ggplot(
         ggplot2::aes(x = dx, y = dy, group = interaction(dist_name, sim_number), color = dist_name)
       ) +
-      ggplot2::geom_line(size = line_size) +
+      ggplot2::geom_line(linewidth = line_size) +
       ggplot2::theme_minimal() +
       ggplot2::labs(
         title = "Density Plot",
@@ -266,7 +266,7 @@ tidy_multi_dist_autoplot <- function(.data, .plot_type = "density", .line_size =
           x = x, y = q, group = interaction(dist_name, sim_number), color = dist_name
         )
       ) +
-      ggplot2::geom_line(size = line_size) +
+      ggplot2::geom_line(linewidth = line_size) +
       ggplot2::theme_minimal() +
       ggplot2::labs(
         title = "Quantile Plot",
@@ -281,7 +281,7 @@ tidy_multi_dist_autoplot <- function(.data, .plot_type = "density", .line_size =
           x = y, group = interaction(dist_name, sim_number), color = dist_name
         )
       ) +
-      ggplot2::stat_ecdf(size = line_size) +
+      ggplot2::stat_ecdf(linewidth = line_size) +
       ggplot2::theme_minimal() +
       ggplot2::labs(
         title = "Probability Plot",
@@ -297,7 +297,7 @@ tidy_multi_dist_autoplot <- function(.data, .plot_type = "density", .line_size =
         )
       ) +
       ggplot2::stat_qq(size = point_size) +
-      ggplot2::stat_qq_line(size = line_size) +
+      ggplot2::stat_qq_line(linewidth = line_size) +
       ggplot2::theme_minimal() +
       ggplot2::labs(
         title = "QQ Plot",
@@ -307,13 +307,13 @@ tidy_multi_dist_autoplot <- function(.data, .plot_type = "density", .line_size =
       ggplot2::theme(legend.position = leg_pos)
   } else if (plot_type == "mcmc") {
     plt <- data_tbl %>%
-      dplyr::group_by(sim_number) %>%
+      dplyr::group_by(sim_number, dist_name) %>%
       dplyr::mutate(cmy = dplyr::cummean(y)) %>%
       dplyr::ungroup() %>%
       ggplot2::ggplot(ggplot2::aes(
-        x = x, y = cmy, group = sim_number, color = sim_number
+        x = x, y = cmy, group = interaction(dist_name, sim_number), color = sim_number
       )) +
-      ggplot2::geom_line() +
+      ggplot2::geom_line(linewidth = line_size) +
       ggplot2::theme_minimal() +
       ggplot2::scale_x_continuous(trans = "log10") +
       ggplot2::labs(